pyproject.toml
[project]
name = "spatial-region-features"
version = "0.1.0"
description = "Research workspace for within-universe spatial features"
requires-python = ">=3.13"
dependencies = [
    # bigwig + region handling (also used by scratch)
    "pybigwig>=0.3.25",
    "pyranges>=0.1.4",
    "gtars>=0.8.0",
    "deeptools>=3.5.6",
    "jupyter>=1.1.1",
    # analysis (plans)
    "numpy>=1.26",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "umap-learn>=0.5",
    # genomic-dict viz project
    "bbconf==0.14.9",
    "bedboss==0.9.6",
    "geniml[ml]>=0.8.4",
    "huggingface_hub>=0.24",
    "polars>=1.0",
    "pybedtools>=0.10",
    "pymemesuite>=0.1.0a4",
    "requests>=2.28",
    "leidenalg>=0.11.0",
    "igraph>=1.0.0",
]
